Fix -Wc++11-narrowing: ash/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2975058
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894744}

+++ ash/accessibility/magnifier/magnifier_glass.cc
@@ -202,7 +202,8 @@ void MagnifierGlass::CreateMagnifierWindow(aura::Window* root_window,
 
   // Create a rounded rect clip, so that only we see a circle of the zoomed
   // content. This circle radius should match that of the drawn border.
-  const gfx::RoundedCornersF kRoundedCorners{params_.radius};
+  const gfx::RoundedCornersF kRoundedCorners{
+      static_cast<float>(params_.radius)};
   zoom_layer_->SetRoundedCornerRadius(kRoundedCorners);
   gfx::Rect clip_rect = window_bounds;
   clip_rect.ClampToCenteredSize(
