+++ ash/public/cpp/external_arc/message_center/arc_notification_content_view.cc
@@ -354,8 +354,8 @@ void ArcNotificationContentView::UpdateControlButtonsVisibility() {
     floating_control_buttons_widget_->Hide();
 }
 
-void ArcNotificationContentView::UpdateCornerRadius(int top_radius,
-                                                    int bottom_radius) {
+void ArcNotificationContentView::UpdateCornerRadius(float top_radius,
+                                                    float bottom_radius) {
   bool force_update =
       top_radius_ != top_radius || bottom_radius_ != bottom_radius;
