+++ ash/wm/splitview/split_view_divider_handler_view.cc
@@ -143,7 +143,7 @@ void SplitViewDividerHandlerView::Refresh(bool is_resizing) {
     selection_animation_->Hide();
 }
 
-void SplitViewDividerHandlerView::UpdateCornerRadius(int radius) {
+void SplitViewDividerHandlerView::UpdateCornerRadius(float radius) {
   layer()->SetRoundedCornerRadius(gfx::RoundedCornersF{radius});
 }
